FoodPortionViewModel (#55)

+++ app/src/main/java/com/iprogrammerr/foodcontroller/model/NutritionalValues.kt
@@ -4,5 +4,5 @@ interface NutritionalValues {
 
     fun calories(): Int
 
-    fun protein(): Int
+    fun protein(): Double
 }
