+++ app/src/main/java/com/iprogrammerr/foodcontroller/model/food/Food.kt
@@ -10,5 +10,5 @@ interface Food {
 
     fun calories(): Int
 
-    fun protein(): Int
+    fun protein(): Double
 }
